// portal/presales_web/src/app/core/enum/roles.enum.ts
/* eslint-disable no-trailing-spaces */
/* eslint-disable @typescript-eslint/naming-convention */
export enum RolesEnum {
    ceo = 'ceo',
    vp = 'vp',
    presales_manager = 'presales_manager',
    presales_consultant = 'presales_consultant',
    executive_manager = 'executive_manager',
    country_manager = 'country_manager',
    sales_manager = 'sales_manager',
    sales = 'sales',
}