// portal/presales_web/src/app/core/interface/opportunityListFilter.interface.ts
/* eslint-disable @typescript-eslint/naming-convention */
export interface IOpportunityListFilter {
   releaseTo: any;
    releaseFrom: any;
    submissionFrom: any;
    submissionTo: any;
    assigne: any[];
    userId: any[];
    assigneId: any[];
    branchId: any[];
    status: any[];
    category: any[];
    rfp_status: any[];
}